add logic to only reboot aapanel if DEBUG is not YES

+++ deployment_scripts/linode-marketplace-aapanel/aapanel-deploy.sh
@@ -153,4 +153,8 @@ function installation_complete {
 # main
 run
 installation_complete
-reboot
+if [[ "${DEBUG}" == "YES" ]]; then
+  :
+else
+  reboot
+fi
